fkdofficial_v6

+++ obr/src/main/java/com/sayone/obr/service/BookServiceImpl.java
@@ -1,11 +1,15 @@
 package com.sayone.obr.service;
 
 import com.sayone.obr.entity.BookEntity;
+import com.sayone.obr.entity.UserEntity;
+import com.sayone.obr.exception.PublisherErrorMessages;
 import com.sayone.obr.repository.BookRepository;
+import com.sayone.obr.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -18,19 +22,21 @@ import java.util.Optional;
 @Service
 @Component
 
-public class BookServiceImpl implements BookService {
+public class BookServiceImpl implements BookService{
 
     @Autowired
     BookRepository bookRepository;
 
-    public BookServiceImpl() {
+    @Autowired
+    UserRepository userRepository;
+
+    public BookServiceImpl(){
 
     }
-
-    //get all books
+//get all books
     @Override
     public List<BookEntity> getBooks() {
-        List<BookEntity> list = (List<BookEntity>) this.bookRepository.findAll();
+        List<BookEntity> list=(List<BookEntity>)this.bookRepository.findAll();
         return list;
     }
 
@@ -41,30 +47,34 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public BookEntity addBook(BookEntity books) {
+    public BookEntity addBook(BookEntity books, Long id) throws Exception {
+        UserEntity userEntity = userRepository.findAllById(id);
+        if (userEntity == null || !Objects.equals(userEntity.getRole(), "publisher")) throw new Exception(PublisherErrorMessages.NO_PUBLISHER_FOUND.getPublisherErrorMessages());
+
+        books.setUid(userRepository.getById(id));
 
         return bookRepository.save(books);
     }
 
 
     @Override
-    public BookEntity updateBook(BookEntity books) {
+    public void deleteBook(Long bId, Long id) throws Exception {
+        UserEntity userEntity = userRepository.findAllById(id);
+        if (userEntity == null || !Objects.equals(userEntity.getRole(), "publisher")) throw new Exception(PublisherErrorMessages.NO_PUBLISHER_FOUND.getPublisherErrorMessages());
 
-        bookRepository.save(books);
-        return books;
+        BookEntity entity = bookRepository.getById(bId);
+        bookRepository.delete(entity);
+
+
     }
 
     @Override
-    public void deleteBook(Long bId) {
-        BookEntity entity = bookRepository.getById(bId);
-        bookRepository.delete(entity);
+    public BookEntity updateBook(BookEntity books, Long id) throws Exception {
+        UserEntity userEntity = userRepository.findAllById(id);
+        if (userEntity == null || !Objects.equals(userEntity.getRole(), "publisher")) throw new Exception(PublisherErrorMessages.NO_PUBLISHER_FOUND.getPublisherErrorMessages());
+        bookRepository.save(books);
+        return books;
     }
-        /*else {
-
-            throw new IllegalStateException("No book found..");
-
-        }*/
-
 
     @Override
     public void uploadBook(MultipartFile file, Long bookId) throws IOException {
@@ -74,7 +84,7 @@ public class BookServiceImpl implements BookService {
         if (optionalUpload.isEmpty()) {
             throw new IllegalStateException("No book found to upload..");
         } else {
-            String path = "/home/meenakshi/updated/obrnew/obr/BookUpload/" + bookId + ".pdf";
+            String path = "/mnt/3160CFF84E83C8AD/SayOne/IDEs/Linux/Java Projects/obr2/obrnew/obr/BookUpload/" + bookId + ".pdf";
             file.transferTo(new File(path));
             findUpload.setBookLink(path);
             bookRepository.save(findUpload);
@@ -84,14 +94,18 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void deleteBookUpload(Long bookId) throws IOException {
+    public void deleteBookUpload(Long bookId, Long id) throws IOException {
+
+        UserEntity userEntity = userRepository.findAllById(id);
+        if (userEntity == null || !Objects.equals(userEntity.getRole(), "publisher")) throw new IOException(PublisherErrorMessages.NO_PUBLISHER_FOUND.getPublisherErrorMessages());
+
         Optional<BookEntity> optionalDelete = bookRepository.findByDeleteArea(bookId);
         BookEntity findDelete = optionalDelete.get();
         String pathCheck = findDelete.getBookLink();
         if (Objects.equals(pathCheck, "") || Objects.equals(pathCheck, "deleted..")) {
             System.out.println("No files to delete");
         } else {
-            Path path = Path.of ("/home/meenakshi/updated/obrnew/obr/BookUpload/" + bookId + ".pdf");
+            Path path = Path.of ("/mnt/3160CFF84E83C8AD/SayOne/IDEs/Linux/Java Projects/obr2/obrnew/obr/BookUpload/" + bookId + ".pdf");
             Files.delete((java.nio.file.Path) path);
             findDelete.setBookLink("deleted..");
             System.out.println("deleted successfully");
